+++ app/lib/screens/home_screen.dart
@@ -9,9 +9,9 @@ import 'package:menstrudel/widgets/period_list_view.dart';
 import 'package:menstrudel/models/period_prediction_result.dart';
 import 'package:menstrudel/utils/period_predictor.dart';
 import 'package:menstrudel/widgets/navigation_bar.dart';
-import 'package:menstrudel/services/period_notifications.dart';
+import 'package:menstrudel/services/notifications/period_notifications.dart';
 import 'package:menstrudel/widgets/dialogs/tampon_reminder_dialog.dart';
-import 'package:menstrudel/services/tampon_notifications.dart';
+import 'package:menstrudel/services/notifications/tampon_notifications.dart';
 
 
 class HomeScreen extends StatefulWidget {
